Version 1.0 final.
This is the version of OpenEval that was submitted to the district
competition. More fixes and changes will be made for version 2.0, which
will be submitted to the state competition. The primary changes that
will be made are: a) Adding support for printing, b) Improving error
handling throughout, and c) Improving comments throughout.